% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gain.R
\name{geom_gain}
\alias{geom_gain}
\title{Geom for gain curve}
\usage{
geom_gain(mapping = NULL, data = NULL, stat = "gain",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aestetic mappings.}

\item{positive_class}{character indicating the "positive" class.}
}
\description{
Geom for gain curve
}
\examples{
library(dplyr)
library(mlbench)
data("BreastCancer")
my_data <- BreastCancer \%>\%
  mutate(Cell.size = as.numeric(Cell.size)) \%>\%
    select(Cell.size, Class)

fit <- glm(Class ~ ., data = my_data, family = "binomial")
my_data$pred <- predict(fit, newdata = my_data)
my_data \%>\%
  ggplot(aes(score = pred, class = Class)) +
  geom_gain(positive_class = "malignant")
}

