% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gain.R
\docType{data}
\name{GeomGain}
\alias{GeomGain}
\title{ggproto class for geom_gain()}
\format{An object of class \code{GeomGain} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomGain
}
\description{
ggproto class for geom_gain()
}
\keyword{datasets}

