% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lift.R
\docType{data}
\name{GeomLift}
\alias{GeomLift}
\title{ggproto class for geom_lift()}
\format{An object of class \code{GeomLift} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomLift
}
\description{
ggproto class for geom_lift()
}
\keyword{datasets}

