% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gain.R
\name{stat_gain}
\alias{stat_gain}
\title{Geom for gain curve}
\usage{
stat_gain(mapping = NULL, data = NULL, geom = "Gain",
  position = "identity", na.rm = FALSE, show.legend = NA, positive_class,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aestetic mappings. Required aesthetics are `score` and
`class`. Further aestetics that can be set are the `fill_color` and
the `fill_alpha`.}

\item{data}{The data to be displayer in this layer.}

\item{positive_class}{character indicating the "positive" class of interest.}
}
\value{
ggplot
}
\description{
Plots the gain curve for evaluating a trained classification
model.
}
\examples{
library(dplyr)
library(mlbench)
data("BreastCancer")
my_data <- BreastCancer \%>\%
  mutate(Cell.size = as.numeric(Cell.size)) \%>\%
    select(Cell.size, Class)

fit <- glm(Class ~ ., data = my_data, family = "binomial")
my_data$pred <- predict(fit, newdata = my_data)
my_data \%>\%
  ggplot(aes(score = pred, class = Class)) +
  geom_gain(positive_class = "malignant")
}

