% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gain.R
\docType{data}
\name{StatGain}
\alias{StatGain}
\title{ggproto class for stat_gain()}
\format{An object of class \code{StatGain} (inherits from \code{Stat}, \code{ggproto}) of length 5.}
\usage{
StatGain
}
\description{
ggproto class for stat_gain()
}
\keyword{datasets}

