% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lift.R
\docType{data}
\name{StatLift}
\alias{StatLift}
\title{ggproto class for stat_lift()}
\format{An object of class \code{StatLift} (inherits from \code{Stat}, \code{ggproto}) of length 5.}
\usage{
StatLift
}
\description{
ggproto class for stat_lift()
}
\keyword{datasets}

