% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gain.R
\name{geom_gain}
\alias{geom_gain}
\title{Geom for lift gain curve}
\usage{
geom_gain(mapping = NULL, data = NULL, stat = "gain",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{formula}{}
}
\description{
Geom for lift gain curve
}
\examples{
fit <- glm(Class ~ Cell.size + Cell.shape, data = BreastCancer, family = "binomial")
fitted <- predict(fit, BreastCancer)
my_data <- data.frame(x = - fitted, y = BreastCancer$Class)
}

